Broken because it's hitting line 64 (can't load with dtypes); suboptimal
It should work but not stating dtypes implies a performance hit that's
probably unacceptable when I turn to the full dataset so this needs to
be fixed soon. It's complaining about strings like "PE", "IN", "IL" in
some column, probably placenames (Peru, Indian/Indiana, Illinois) but
I'm trying to narrow down which column.